Fixing reason for modulemanager crash on target.
using vf-cpp the modulemanager would crash on the target.
The reason is that the = operator is not returning an object.

This ends in a segmentation fault and kills the thread with:
terminate called without an active exception

The reason is that assigning a value to veincompproxy with = operator
did not retrun the correct object or any object at all.

Signed-off-by: bhamacher <b.hamacher@zera.de>