Changed readonly to direction in VeinModuleComponent.
Until now it was possible to set readonly to true.
It was not possible to write data from outside the entityin this case.

now:
inOut: equal to  readonly=false
out: equal to readonly=true
in: only extern enities can write data to the component.

This behavior is usefull, when components act as interfaces to some kind of black box (model based software development).

Signed-off-by: bhamacher <b.hamacher@zera.de>