Added support for VEIN dissassemble
Until now it was only possible to create components and enities.
Now we can remove components with the VeinSharedComp::removeComponent call.
(cVeinModuleComponent is not working with removeComponent because there are still old API
dependecies. We will solve those in future. Use VeinModuleEntitie::removeCompontent!).
VeinSharedComponent functions check if the component is equal nullptr and
prevent access in this case. This keeps using a component always safe.

Deleting a VeinModuleEntitie object removes the entity from the eventHandler.
All Components and rpcs being part of this entitie will be deleted.

Deleting any kind of vf-cpp object will cause an event signing out those in VEIN.

To manage references QSharedPointer and QWeakPointer is used. Therefore all operation are threadsafe.

Attention! Some of those feature needed changes in vf-event. This commit will only
work with those. Search for "VeinDepDisAs" to find the fitting commit.

Signed-off-by: bhamacher <b.hamacher@zera.de>